+++ src/main/kotlin/dsm/pick2024/domain/bug/AddBugService.kt
@@ -2,7 +2,7 @@ package dsm.pick2024.domain.bug
 
 import dsm.pick2024.domain.bug.presentation.dto.request.BugRequest
 import dsm.pick2024.domain.user.facade.UserFacade
-import dsm.pick2024.global.discord.DiscordMessage
+import dsm.pick2024.infrastructure.feign.client.dto.request.DiscordMessageRequest
 import dsm.pick2024.infrastructure.feign.client.DiscordBugClient
 import dsm.pick2024.infrastructure.s3.FileUtil
 import dsm.pick2024.infrastructure.s3.PathList
@@ -17,10 +17,10 @@ class AddBugService(
 ) {
     @Async
     fun bugAlarm(request: BugRequest) {
-        val message = DiscordMessage(
+        val message = DiscordMessageRequest(
             content = "# 🚨 버그 제보",
             embeds = listOf(
-                DiscordMessage.Embed(
+                DiscordMessageRequest.Embed(
                     title = "ℹ️ 버그 정보",
                     description = """
                         ### 🕖 버그 제목
