+++ src/main/kotlin/dsm/pick2024/infrastructure/feign/NeisMealFeignClientService.kt
@@ -4,7 +4,7 @@ import com.google.gson.Gson
 import dsm.pick2024.domain.meal.entity.MealJpaEntity
 import dsm.pick2024.infrastructure.feign.client.NeisFeignClient
 import dsm.pick2024.infrastructure.feign.client.property.NeisFeignClientRequestProperty
-import dsm.pick2024.infrastructure.feign.client.response.NeisFeignClientMealServiceDietInfoResponse
+import dsm.pick2024.infrastructure.feign.client.dto.response.NeisFeignClientMealServiceDietInfoResponse
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import java.time.LocalDate
