+++ src/main/kotlin/dsm/pick2024/infrastructure/feign/client/DiscordProdClient.kt
@@ -1,6 +1,6 @@
 package dsm.pick2024.infrastructure.feign.client
 
-import dsm.pick2024.global.discord.DiscordMessage
+import dsm.pick2024.infrastructure.feign.client.dto.request.DiscordMessageRequest
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -8,5 +8,5 @@ import org.springframework.web.bind.annotation.RequestBody
 @FeignClient(name = "discord-bug-prod-client", url = "\${discord.webhook.prod}")
 interface DiscordProdClient {
     @PostMapping
-    fun prodSendAlarm(@RequestBody message: DiscordMessage)
+    fun prodSendAlarm(@RequestBody message: DiscordMessageRequest)
 }
