+++ src/main/kotlin/dsm/pick2024/infrastructure/feign/client/XquareFeignClient.kt
@@ -1,7 +1,7 @@
 package dsm.pick2024.infrastructure.feign.client
 
-import dsm.pick2024.infrastructure.feign.client.response.XquareResponse
-import dsm.pick2024.infrastructure.feign.client.response.XquareUserAllResponse
+import dsm.pick2024.infrastructure.feign.client.dto.response.XquareResponse
+import dsm.pick2024.infrastructure.feign.client.dto.response.XquareUserAllResponse
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
