+++ src/main/kotlin/dsm/pick2024/infrastructure/feign/client/dto/response/XquareResponse.kt
@@ -1,4 +1,4 @@
-package dsm.pick2024.infrastructure.feign.client.response
+package dsm.pick2024.infrastructure.feign.client.dto.response
 
 import java.time.LocalDate
 import java.util.*
